SAA-625: Jest updates & tests
